+++ docs/source/changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.12.12 (2026-XX-XX)
-- Fix Polars dataframe upload fallbacks for MSSQL and Snowflake backends
+- Fix dataframe upload fallbacks for MSSQL and Snowflake backends
 
 ## 0.12.11 (2026-01-23)
 - Feat: simplify accessing tables in ParquetTableStore with polars